app/requirements.txt
python-dotenv
fastapi
llama-index-llms-groq
llama-index-experimental
matplotlib
numpy
pandas
pydantic
requests
scikit-learn
streamlit
uvicorn
ydata-profiling
